Refactor PR preview links to match consistent main preview experience
