+++ frontend/src/app/(dashboard)/previews/[id]/page.test.tsx
@@ -64,7 +64,7 @@ describe("PreviewLandingPage launch mode", () => {
     expect(screen.getByText("529975ce1faa")).toBeInTheDocument();
     expect(screen.getByText("Opening when ready")).toBeInTheDocument();
     expect(screen.getByText("This preview will open automatically when it is ready.")).toBeInTheDocument();
-    expect(screen.getByText("Start services")).toBeInTheDocument();
+    expect(screen.getAllByText("Start Services").length).toBeGreaterThan(0);
   });
 
   it("waits for bootstrap completion before navigating to the preview origin", async () => {
@@ -559,5 +559,36 @@ describe("PreviewLandingPage status mode", () => {
     ).toBeInTheDocument();
     expect(screen.getByRole("button", { name: "Start preview" })).toBeInTheDocument();
     expect(screen.getByRole("button", { name: "Preview actions" })).toBeInTheDocument();
+  });
+
+  it("warns when new commits are available on the branch preview", async () => {
+    searchParams = new URLSearchParams("");
+
+    server.use(
+      http.get("*/api/v1/previews/target-1", () =>
+        HttpResponse.json({
+          data: {
+            target_id: "target-1",
+            preview_id: "prev-1",
+            repository_id: "repo-1",
+            repository_full_name: "acme/web",
+            branch: "feature/preview",
+            commit_sha: "529975ce1faa2961ef3f23abde2418bf561116d9",
+            latest_commit_sha: "def456abc7890000000000000000000000000000",
+            new_commits_available: true,
+            source_type: "pull_request",
+            status: "ready",
+            current_phase: "ready",
+            stable_url: "https://143.dev/previews/target-1",
+            preview_url: "https://target-1.preview.143.dev",
+          },
+        }),
+      ),
+    );
+
+    renderLaunchPage();
+
+    expect(await screen.findByText("New commits available")).toBeInTheDocument();
+    expect(screen.getByText("Latest: def456abc789")).toBeInTheDocument();
   });
 });
